[NO-TICKET] Fix Ruby jobs failing due to library rename
**What does this PR do?**

This PR changes prof-correctness to match the library rename for version
2 of the Ruby library.

* The gem is now called `datadog` instead of `ddtrace`
* The Ruby profiler command wrapper is called `ddprofrb` instead of
  `ddtracerb`

**Motivation:**

Because prof-correctness was running using the latest version of the
Ruby library from the master branch, it broke when the 2.0 development
branch (which included the library rename) was merged to master in
https://github.com/DataDog/dd-trace-rb/pull/3421 .

**Additional Notes:**

I don't expect customers to run into the same issue, since you
wouldn't normally be sourcing the library from master directly
(that's really bleeding edge!).

**How to test the change?**

Validate that prof-correctlness Ruby CI jobs are passing again.