[NO-TICKET] Minor: Use `wait_until_running` for ruby tests
**What does this PR do?**

The Ruby profiler runs in a background test, and so often in our tests
we need a way to make sure the profiler is already running before we
actually start the test (otherwise we may miss some of the data which
may lead to flaky failures).

I got tired of this boilerplate and introduced an actual API to do this
in the profiler -- `Datadog::Profiling.wait_until_running`.

This PR replaces our previous hacky boilerplate to ensure the profiler
was running with this new API.

**Motivation:**

Reduce boilerplate in tests.

**Additional Notes:**

N/A

**How to test the change?**

Validate that Ruby scenarios still pass.