Fix incorrectly removing extension
As I was preparing to merge
https://github.com/DataDog/prof-correctness/pull/39 I noticed that CI
had not failed for Ruby prior to version 2.2.0 of the Ruby library being
released.

(Specifically: it had failed, but not for Ruby, which is why I had not
spotted it before).

The whole point of this test was that it was _supposed_ to be failing
prior to 2.2.0 being released.

Under closer examination, what happened was that I had tested with the
1.x branch prior to opening the PR (and all my examples in the PR
description are from 1.x), and of course I missed that the test setup
was not correct for triggering the issue on 2.x due to the library
getting renamed.

I've now fixed this, confirmed it's broken with versions < 2.2.0 of the
Ruby library, and confirmed it passes for the right reason on >= 2.2.0.

This PR is thus ready to be merged :)