[PROF-13273] Add workaround for bug causing tests to fail
**What does this PR do?**

This PR adds a workaround to
https://github.com/DataDog/dd-trace-rb/issues/5137 which is making
the `ruby_extension_dir_and_rpath` test to fail: it skips testing
with broken `2.23.0` library version.

This is similar to the workaround we added in profiling-backend:
https://github.com/DataDog/profiling-backend/pull/7754 .

By avoiding the exact release where the regression shipped we'll
take Ruby out of the red, and once the next release comes in
(2.24.0 or 2.23.1) it'll get picked up **automatically** again.

**Motivation:**

Get prof-correctness for Ruby back to green.

**Additional Notes:**

The test was failing with:

```
   correctness_test.go:167: Error running the test docker run -v /home/runner/work/prof-correctness/prof-correctness/data/ruby_heap_highload-3461119285:/app/data:rw -u 1001:1001 --security-opt seccomp=unconfined -e DD_SERVICE=prof-correctness-ruby_heap_highload test-app:latest - /usr/local/lib/ruby/3.3.0/bundler/shared_helpers.rb:107:in `rescue in filesystem_access': There was an error while trying to write to `/app/bin`. It is likely that you need to grant write permissions for that path. (Bundler::PermissionError)
```

**How to test the change?**

Validate Ruby is back to green!